+++ src/components/Hero.tsx
@@ -2,6 +2,7 @@ import { useRef, useState } from "react";
 import { useDebounce, useDidUpdate, useKeys } from "rooks";
 import { FaRegDotCircle } from "react-icons/fa";
 import { AiOutlineStar } from "react-icons/ai";
+import Avatar from "./Avatar";
 import { fetchRecommendations } from "../lib/supabase";
 import humanizeNumber from "../lib/humanizeNumber";
 import { getAvatarLink } from "../lib/github";
@@ -91,7 +92,7 @@ const Hero = () => {
               <p className="text-gray-500 text-sm font-semibold">Repository</p>
             </div>
 
-            {fetchedData.map(({ full_name, name, description, issues, stars }) => (
+            {fetchedData.map(({ full_name, name, description, issues, stars, contributions }) => (
               <a
                 key={full_name}
                 href={`https://app.opensauced.pizza/repos/${full_name}`}
@@ -120,9 +121,19 @@ const Hero = () => {
 
                     <div className="flex justify-between mt-[8px]">
                       <div className="flex gap-x-[5px]">
-                        <div className="w-[20px] h-[20px] rounded-full">
-                          {/* todos: add contributors avator here */}
-                        </div>
+                        {contributions.slice(0, 3).map(({ contributor, last_merged_at }) => (
+                          <div
+                            key={`search-${full_name}-${contributor}`}
+                            className="w-[20px] h-[20px] rounded-full"
+                          >
+                            <div className="flex items-center overflow-hidden rounded-full w-8 h-8">
+                              <Avatar
+                                contributor={contributor}
+                                lastPr={last_merged_at}
+                              />
+                            </div>
+                          </div>
+                        ))}
                       </div>
 
                       <div className="flex gap-x-[6px]">